+++ __tests__/html/chatHistory.restore.sent.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('when restoring chat history', () => {
+  test('outgoing activities should be marked as sent', () => runHTML('chatHistory.restore.sent.html'));
+});